* validating user * saving and retrieving of variants

+++ db/app_migrations/008_create_tags.rb
@@ -7,7 +7,7 @@ class CreateTags < Sequel::Migration
       String :name, size: 255, unique: true # Original name, for now the actual name of the tag, in light of i18n it will be the reference moniker for translations and the tag_id will be used to map translations
 
       String :description
-      Integer :taxonomy # domain value
+      Integer :taxonomy # domain value. paper format, suitable target age, braille type etc
       DateTime :created_at
 
       index :name
@@ -16,6 +16,8 @@ class CreateTags < Sequel::Migration
   end
 
   def down
-    drop_table :tags
+    if @db.table_exists?(:tags)
+      drop_table :tags
+    end
   end
 end
